requirements.txt
asgiref==3.9.1
attrs==25.4.0
awsebcli==3.25.3
bcrypt==5.0.0
blessed==1.25.0
botocore==1.42.17
cement==2.10.14
certifi==2025.7.14
cffi==2.0.0
charset-normalizer==3.4.2
colorama==0.4.6
cryptography==45.0.5
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.3.1
dj-database-url==3.0.1
Django==5.2.4
django-jazzmin==3.0.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-jwt==1.0.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-nested-routers==0.93.5
drf-spectacular==0.29.0
fabric==3.2.2
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
invoke==2.2.1
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
oauthlib==3.3.1
packaging==24.2
paramiko==4.0.0
pathspec==0.12.1
pillow==11.3.0
pycparser==2.22
PyJWT==2.10.1
PyNaCl==1.6.1
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.4
requests-oauthlib==2.0.0
rpds-py==0.30.0
semantic-version==2.10.0
setuptools==80.9.0
six==1.17.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
termcolor==2.5.0
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==1.26.20
wcwidth==0.2.14
whitenoise==6.11.0
wrapt==2.0.1
